refactor(client-angular): Replace JS confirms with MatDialogs

+++ client/crm-example-client-angular/src/app/feature/task-detail/task-detail.component.ts
@@ -8,7 +8,9 @@ import { LayoutComponent } from '../../shared/layout/layout.component'
 import { LineBreakPipe } from '../../shared/line-break.pipe'
 import { ActivatedRoute, Router, RouterLink } from '@angular/router'
 import { TaskFormComponent } from '../../shared/task-form/task-form.component'
-import { switchMap, tap } from 'rxjs'
+import { of, switchMap, tap } from 'rxjs'
+import { MatDialog } from '@angular/material/dialog'
+import { DialogComponent, IDialogData } from '../../shared/dialog/dialog.component'
 
 @Component({
   selector: 'app-task-detail',
@@ -28,6 +30,7 @@ import { switchMap, tap } from 'rxjs'
 })
 export class TaskDetailComponent implements OnInit {
   private activatedRoute = inject(ActivatedRoute)
+  private dialog = inject(MatDialog)
   private router = inject(Router)
   private tasksService = inject(TasksService)
 
@@ -68,9 +71,30 @@ export class TaskDetailComponent implements OnInit {
   }
 
   protected deleteTask() {
-    if (!confirm('Are you sure you want to delete this task?\n This is irreversible.')) return
 
-    this.tasksService.deleteTask(this.taskId)
-      .subscribe(() => this.router.navigate(['/tasks']))
+    this.dialog.open(DialogComponent, { data: {
+      title: 'Confirm Delete Task',
+      contents: [
+        'Are you sure you want to delete this task?',
+        'This is irreversible.',
+      ],
+      actions: [
+        { text: 'Cancel' },
+        { value: true, text: 'Delete Confirmed' },
+      ],
+    } as IDialogData })
+    .afterClosed()
+    .pipe(
+      switchMap(confirmed => confirmed ? this.tasksService.deleteTask(this.taskId) : of(null)),
+      switchMap(response => {
+        if (response) {
+          return this.dialog.open(DialogComponent, { data: {
+            title: 'Task Deleted',
+            actions: [{ value: true, text: 'Confirm' }],
+          } as IDialogData }).afterClosed()
+        }
+        return of(null)
+      }),
+    ).subscribe(confirmed => confirmed ? this.router.navigate(['/tasks']) : null)
   }
 }
